Revert "death and reviving"
This reverts commit 20127a314730bb003118b16e0f4dfd8aa02b9b51.

+++ Assets/Scripts/Character/Player/PlayerUIHudManager.cs
@@ -13,13 +13,13 @@ public class PlayerUIHudManager : MonoBehaviour
         staminaBar.gameObject.SetActive(true);
     }
 
-    public void SetNewHealthValue(int oldValue, int newValue)
+    public void SetNewHealthValue(float oldValue, float newValue)
     {
         healthBar.SetStat(Mathf.RoundToInt(newValue));
 
     }
 
-    public void SetMaxHealthValue(int maxHealth)
+    public void SetMaxHealthValue(float maxHealth)
     {
         healthBar.SetMaxStat(maxHealth);
     }
